+++ src/pipeline_telemetry/decorator.py
@@ -15,7 +15,7 @@ def add_telemetry(telemetry_params: dict) -> object:
 
     Args:
         telemetry_params (dict:
-            - process_name (str): free format process name
+            - source_name (str): free format process name
             - process_type (ProcessType): process type definition
             - telemetry_rules (dict): telemetry rules definition dict
             - storage_class (AbstractTelemetryStorage, optional):
@@ -30,9 +30,12 @@ def add_telemetry(telemetry_params: dict) -> object:
             """
             if (not hasattr(self, "_telemetry")) or (not self._telemetry):
                 self._telemetry = Telemetry(**telemetry_params)
-            result = method(self, *args, **kwargs)
-            self._telemetry.save_and_close()
-            self._telemetry = None
+                result = method(self, *args, **kwargs)
+                self._telemetry.save_and_close()
+                self._telemetry = None
+            else:
+                result = method(self, *args, **kwargs)
+
             return result
 
         return wrapped_method
@@ -48,7 +51,7 @@ def add_mongo_telemetry(telemetry_params: dict) -> object:
 
     Args:
         telemetry_params (dict:
-            - process_name (str): free format process name
+            - source_name (str): free format process name
             - process_type (ProcessType): process type definition
             - telemetry_rules (dict): telemetry rules definition dict
     """
@@ -60,9 +63,11 @@ def add_mongo_telemetry(telemetry_params: dict) -> object:
             """
             if (not hasattr(self, "_telemetry")) or (not self._telemetry):
                 self._telemetry = mongo_telemetry(**telemetry_params)
-            result = method(self, *args, **kwargs)
-            self._telemetry.save_and_close()
-            self._telemetry = None
+                result = method(self, *args, **kwargs)
+                self._telemetry.save_and_close()
+                self._telemetry = None
+            else:
+                result = method(self, *args, **kwargs)
             return result
 
         return wrapped_method