+++ src/pipeline_telemetry/main.py
@@ -25,7 +25,9 @@ from pipeline_telemetry.validators.dict_validator import DictValidator
 BASE_COUNT_KEY = "base_counter"
 ERRORS_KEY = "errors"
 FAIL_COUNT_KEY = "fail_counter"
-PROCESS_NAME = "process_name"
+SOURCE_NAME_KEY = "source_name"
+CATEGORY_KEY = "category"
+SUB_CATEGORY_KEY = "sub_category"
 PROCESS_TYPE_KEY = "process_type"
 START_TIME = "start_date_time"
 RUN_TIME = "run_time_in_seconds"
@@ -62,7 +64,7 @@ class Telemetry:
     a database provide via a storage class.
 
     args:
-        - process_name (str):
+        - source_name (str):
             Process name for which telemetry is made for example: `GET-WEATHER`
             Process names can be choosen freely and should ne unique for a
             single process as they are used to collect all telemetry data for a
@@ -84,7 +86,9 @@ class Telemetry:
         - add_process_types: add custom process types
 
     properties:
-        - process_name
+        - category
+        - sub_category
+        - source_name
         - process_type
         - sub_process_types
 
@@ -102,14 +106,18 @@ class Telemetry:
 
     def __init__(
         self,
-        process_name: str,
+        category: str,
+        sub_category: str,
+        source_name: str,
         process_type: ProcessType,
         telemetry_rules: dict = None,
         storage_class: AbstractTelemetryStorage = None,
     ):
         self._set_process_type(process_type)
         self._telemetry = {
-            PROCESS_NAME: process_name,
+            CATEGORY_KEY: category,
+            SUB_CATEGORY_KEY: sub_category,
+            SOURCE_NAME_KEY: source_name,
             PROCESS_TYPE_KEY: self._process_type.name,
             START_TIME: datetime.now(),
         }
@@ -158,9 +166,9 @@ class Telemetry:
         return storage_class()
 
     @property
-    def process_name(self) -> str:
-        """Process_name property."""
-        return self._telemetry.get(PROCESS_NAME)
+    def source_name(self) -> str:
+        """Source_name property."""
+        return self._telemetry.get(SOURCE_NAME_KEY)
 
     @property
     def telemetry(self) -> dict:
@@ -415,11 +423,17 @@ class Telemetry:
 
 
 def mongo_telemetry(
-    process_name: str, process_type: ProcessType, telemetry_rules: dict
+    category: str,
+    sub_category: str,
+    source_name: str,
+    process_type: ProcessType,
+    telemetry_rules: dict,
 ) -> Telemetry:
     """Factory method to create Telemetry instance with MongoStorage class."""
     return Telemetry(
-        process_name=process_name,
+        category=category,
+        sub_category=sub_category,
+        source_name=source_name,
         process_type=process_type,
         telemetry_rules=telemetry_rules,
         storage_class=TelemetryMongoStorage,